Add ISV product ID and ISV SVN fields to manifest (#502)
ISV product ID and ISV SVN fields are defined and signed in SIGSTRUCT.
In Gramine, they are defined in the manifest and then used on enclave
signing in enclave build flow. Add the option to configure them in
OpenFL manifest with new flags to the makefile: SGX_ISVPRODID and
SGX_ISVSVN. Their default values are 0, to maintain backwards
compatibility.

Currently ISV SVN is configurable in build time, but this is not
necessarily the desired behavior. Required behavior is open for
discussion and should be updated once a decision is made.