Support configurable enclave size in graminize
Currently OpenFL enclave size is hardcoded to 16G. Add option to
configure it in build time:
Add the "-e" argument to "fx workspace graminize", to allow setting
the size of the enclave that will be used in the image. This setting
propagates as a Docker build argument to the makefile in openfl-gramine,
which eventually becomes an argument in the generated manifest.
Default size is 16G, to maintain backwards compatibility.

Fixes #568

Signed-off-by: Lavi, Nir <nir.lavi@intel.com>