Print CSR hash in generate-cert-request (#692)
Currently, there is no real way to verify CSR properly when using fx
command line: since generate-cert-request does not print the CSR hash,
it is not really possible to know in certify that the CSR being signed
is the correct one. To mitigate this, add a print with the CSR hash in
generate-cert-request. User is then expected to compare the hashes and
approve the CSR only if they match

Signed-off-by: Lavi, Nir <nir.lavi@intel.com>